Add SafeSerializer that performs validation & apply encryption.
